shadow: skip copying extended attributes in useradd when enable SELinux
Issue: LIN1018-1809

The patch 0001-useradd-copy-extended-attributes-of-home.patch introduced
a runtime issue when enable SELinux feature. With this patch, the
directory /home/user will get the extended attributes of /etc/skel.
However, when enable SELinux feature, the SELinux object lable for
/etc/skel is etc_t which also copy to /home/user. This will cause the
users can not write their home directory because the SELinux object
lable for /home/user should be user_home_dir_t.

There is a discussion in oe-core mailing list:
http://lists.openembedded.org/pipermail/openembedded-core/2018-January/146409.html

This patch is mainly used for SMACK, so we can skip copying the extended
attributes when enable SELinux feature.

(LOCAL REV: NOT UPSTREAM) -- will send to oe-core on 20180820

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>