grub-efi-native: workaround for fedora 29
Issue: LIN1018-3095

grub-efi-native fails to run do_configure on Fedora 29 host:

| checking for options to compile assembly... \nRunning gcc -c -o /dev/null -g -Wall
|  -W -DGRUB_MACHINE_EFI=1 -DGRUB_MACHINE=X86_64_EFI ../grub-2.02/asm-tests/i386.S
| Assembler messages:
| Fatal error: The input and output files must be distinct

because the host 'as' in binutils-3.31 breaks if using two filesystems.
See:

https://sourceware.org/git/gitweb.cgi?p=binutils-gdb.git;a=commit;h=2a50366

Work-around this issue by forcing the configure time assembler check
to use a temp file in the src dir.

(LOCAL REV: NOT UPSTREAM) -- Send to oe-core on 20181219

Signed-off-by: Kai Kang <kai.kang@windriver.com>