linux-libc-headers: Fix build failure by using fixed input and output files instead of pipe
Issue: LIN1018-3188

This is a workaround for the following possible build failure.

*** Compiler lacks asm-goto support.. Stop.

When building linux-libc-headers we need to use binutils on build machine.
binutils v2.31 introduces a bug that could cause scripts/gcc-goto.sh to fail
when running in an environment where /tmp is rarely used, e.g. in docker.

(LOCAL REV: NOT UPSTREAM) -- Temporary for LTS18, will be sent to oe-core if OK

Signed-off-by: He Zhe <zhe.he@windriver.com>