oe-core: fix missing chkconfig header in init script
Issue: LIN1018-3287

chkconfig needs a specific header to work properly,
which indicates the initial start and stop order.
Or it will mess up the priority of init scripts while
enable or disable service. So it will be necessary to
add this header into init script which supports
chkconfig.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20181204

Signed-off-by: Rui Wang <rui.wang@windriver.com>