tzdata: Install everything by default
Issue: LIN1018-4196

tzdata is converted to an empty meta package which pulls in all
subpackages. The subpackages are defined in a TZ_PACKAGES variable so
that we don't have to repeat ourselves.

The timezones and conffiles which were in the tzdata package are moved
to a new 'tzdata-core' package.

Signed-off-by: Paul Barker <paul@betafive.co.uk>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
(master rev: 2af4d6eb2526d60b26bc5128068541ff3350fb58)
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>