init-install-efi.sh: don't assume the fixed location of efi loader
Issue: US132888

Currently, it is true that efi loader is in the same device as the
rootfs image. But the script doesn't work when efi loader and rootfs
image are in different live devices. This change makes the script work
in this situation.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20190327

Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>